cloud_functions/webhook/requirements.txt
functions-framework==3.*
google-api-python-client
textract
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl